Database complete with all keys in place
Database has been completed with all current keys in place. Due to scope concerns I have opted to omit the Rental dispute in the interest of first getting key requirements working. If we have enough time after getting the key requirements fulfilled we will add that table and work on that system later.