User Roles Implimented on the Login
This update has working user roles. The login page will now detect if the propert user role is present upon login. Everyone can see the customer panel. If you have a user role of Property Manager it will then show the Property Manager Panel.

Made a few tweaks to the database to make some primarty IDs auto incriment. I forgot to do that on the initial setup for some tables so now that is fixed.