House Keeping Update
Changed the project to include directories for each specific dashboard. Cleaned up the main directory and moved all pages to the specific directories for their purpose.  Finally I updated the links on the login pages and moved pages that may be cut to a specific folder till we determine that those features will be perminately removed.