3 Page Update!
Got the AdminDash pages of ManageStates.aspx, and ManageCities.aspx working. The pages allow for the viewing of the data in the table as well as being able to add new data.

The other page I got half functioning was the CustomerDash page of UserInfo.aspx The page now properly populates with the users data on load. All thats left is to allow them to edit the data.