Setup Static controls
static controls have been setup but the code behind still has to be modified to use them.

I think that if we wanted to try to get the buttons working through code though we would need to start by finding out how to assign a click event to a button that hasn't been created yet. I'm not sure thats possible though. I keep thinking that if we could set a click event in the load() then maybe it would work but since their is no button yet I don't know how we would interface with it.